Add re-worked regression tests.
These are the regression tests used during the development of Postgres 9.2's
pg_stat_statements normalization, repurposed for plan normalization. It is
generally expected that assuming that there is a 1:1 correspondance between
query plans are queries (which, of course, there is not) pg_stat_plans will for
the most part server as a functional replacement for the 9.2 module on 9.1.