Document limitations, add summary view.
I've documented some of the limitations with the plan-hashing approach that
previously caused some tests to fail (essentially, preprocess_expression would
preprocess expressions that didin't reference some relation during planning, and
it is not possible to work backwards from the Const within the finished plan). A
few of the relevant tests have been removed or re-written to reference a
relation, but probably not all of them.

Also, added a view that uses regular expressions to produce a summary of plan
execution costs at the query granularity. Of course, the normalisation is only
as good as the regular expression, but it is expected to be sufficient much of
the time.