Fix some more test cases.
These didn't pass due to simplification of constant expressions. They were
changed to refer to a Var, so that this doesn't happen (I could just as easily
have written the test cases this way when I originally did so for query-tree
based normalization).

This is a documented limitation of pg_stat_plans, and not a particularly
pressing one.