Add search_path fingerprinting.
The idea is to fingerprint the search_path setting of the user during query
execution, allowing us to recognize that as the likely reason for planid
differences when we go to explain query text. We only error when the planids
actually fully differ, because two different search_path settings may be similar
enough to have the query actually refer to the same relation OIDs.

The mechanism throuh which this is done is a kludge, but I see no reasonable
alternative.

In passing, document all GUCs that the module provides. More extensively
describe limitations with our approach on explaining entries, too.