Track start-up and total costs for plans.
Although also available from EXPLAIN output, this is still useful. It may not
even be possible to get EXPLAIN text for the entry if the original query string
no longer produces the same plan. This is precisely when that information will
be most useful, as it may be used by the user to deduce the point at which the
planner prefers a new plan.

Per suggestion from Stephen Frost.