Tweak pg_stat_plans_explain() semantics for the convenience of tools.
Previously, we warned in the event of a query text that produced a different
plan than that stored for the entry, while still returning the new plan, with
another warning inline, in the explain text. This warning is likely to hamper
third party parsing tools. Moreover, this behavior wasn't particularly
intuitive, since the user asked for the plan text for a particular entry/plan,
and not a particular query text. Just return NULL instead, while continuing to
raise a WARNING (or, if search_paths differ, ERROR) as before.