Tweak the query_valid column to indicate query explainability generally.
The query_valid column has been renamed to query_explainable, and has been
broadened in what it indicates; it now indicates if an entry's query text is
expected to be explainable (or rather, not explainable) for any reason, not just
whether or not the text is currently considered valid.

This has been done for the convenience of third-party analysis tools, that might
otherwise be forced to continually speculatively run pg_stat_plans_explain() on
the same entry. With an application that uses prepared queries selectively, such
as an application with some number of stored functions, the third party analysis
tool could produce an unreasonable level of server log noise. It seems unlikely
that the distinction between query text validity (i.e that a query text is
thought to currently produce the plan for its entry when explained) and general
entry text explainability is a useful enough one to expose to the user.