Add STAT_PLANS_DEBUG infrastructure
This infrastructure can be used to debug problems with pg_stat_plans
fingerprinting logic. When the macro STATS_PLAN_DEBUG is defined, Postgres will
print a low-level representation of each plan tree that is fingerprinted, as it
is fingerprinted.

Add documentation on how to debug problems with pg_stat_plans' fingerprinting
logic, including using this new infrastructure.