Fix backend crash in case an empty search_path was set.
This could happen in case a database session has set a search_path
to a schema not accessible by the current role for example.
This results in an empty activeSearchPath list, which is returned by
fetch_search_path() as an empty list. Access in get_search_path_xor()
then causes a crash.

The best solution to this seems to also include any implicit search_path
members, too. This also seems to be a more correct definition in the sense
of a "stable" view when evaluating the plans later.